refactor(#3479): extract panel-lifecycle rendering from turn_bridge/mod.rs
Verbatim-move the task/session-panel line rendering + active-placeholder-card
cluster out of turn_bridge/mod.rs into a new `panel_lifecycle.rs` leaf:
first_request_line, child_progress_line, refresh_session_panel_line_from_lifecycle,
refresh_task_panel_line_from_dispatch, ensure_active_placeholder_card, plus the
private TaskPanelDispatchMetadata struct + load_task_panel_dispatch_metadata.

Behavior-preserving: only visibility prefixes (pub(super)) and `super::` ->
`super::super::` depth were adjusted; logic is byte-identical. turn_bridge/mod.rs
6448 -> 6239 prod LoC; giant ratchet lowered to lock the win.

codex adversarial review CLEAN (r1 flagged only a stale change-surfaces.md count,
now updated to 6239).
